+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added code editor for primitivesLibrary field
+- ConstructorSchemes: Added loading svg icons from server
 
 ### Changed